+++ example/example.dart
@@ -7,21 +7,21 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
+  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Count Down',
+      title: 'Flutter Demo',
       theme: ThemeData(
         primarySwatch: Colors.blue,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-      home: MyHomePage(title: 'Flutter Count Down'),
+      home: MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
 
   final String title;
 
@@ -30,17 +30,8 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  String countDown = 'Loading';
-
   @override
   Widget build(BuildContext context) {
-    //List of Date to countDown
-    List<DateTime> s = [
-      DateTime.utc(2022),
-      DateTime.utc(2021),
-      DateTime(2021, 2, 7)
-    ];
-
     return Scaffold(
       appBar: AppBar(
         title: Text(widget.title),
@@ -50,27 +41,17 @@ class _MyHomePageState extends State<MyHomePage> {
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
             Text(
-              'You have pushed the button this many times:',
+              'Countdown Timer:',
             ),
-            Expanded(
-              child: ListView(
-                children: List.generate(3, (index) {
-                  countDown = CountDown()
-                      .timeLeft(s[index], 'Time Ended', longDateName: true);
-                  return Text(
-                    countDown,
-                    style: Theme.of(context).textTheme.headline4,
-                  );
-                }),
-              ),
+            CountDownText(
+              due: DateTime.utc(2050),
+              finishedText: "Done",
+              showLabel: true,
+              longDateName: true,
+              style: TextStyle(color: Colors.blue),
             )
           ],
         ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {},
-        tooltip: 'Increment',
-        child: Icon(Icons.add),
       ),
     );
   }